pipeline/requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
openai>=1.30.0
simhash>=2.1.0
python-dateutil>=2.9.0
pytz>=2024.1
python-dotenv>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0
